Forgot password refactor
